# tests/baseline/openapi_app_snapshot.json
{
  "components": {
    "schemas": {
...
      "CreateSubscriptionBody": {
        "description": "Request body for creating a subscription under a strategy template.\n\n``symbol`` is composite ``{code}:{exchange}`` (e.g. ``BTCUSDT:BINANCE``).",
        "properties": {
          "interval": {
            "title": "Interval",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "interval"
        ],
        "title": "CreateSubscriptionBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Algorithmic trading platform — API, SPA serving, and trading runtime",
    "title": "PocketQuant",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/backtest/strategy/{strategy_id}": {
      "get": {
        "operationId": "list_backtests_api_v1_backtest_strategy__strategy_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_failed",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Failed",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Backtests Api V1 Backtest Strategy  Strategy Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Backtests",
        "tags": [
          "backtest"
        ]
      }
    },
...
    "/api/v1/strategies/{strategy_code}": {
      "delete": {
        "operationId": "delete_strategy_by_code_api_v1_strategies__strategy_code__delete",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_code",
            "required": true,
            "schema": {
              "title": "Strategy Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Strategy By Code",
        "tags": [
          "strategies"
        ]
      },
      "get": {
        "operationId": "get_strategy_template_api_v1_strategies__strategy_code__get",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_code",
            "required": true,
            "schema": {
              "title": "Strategy Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Strategy Template Api V1 Strategies  Strategy Code  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Strategy Template",
        "tags": [
          "strategies"
        ]
      }
    },
...
    "/api/v1/strategies/{strategy_code}/subscriptions": {
      "post": {
        "description": "Create a subscription that binds a strategy template to (symbol, interval).\n\n``symbol`` must be composite ``{code}:{exchange}``.",
        "operationId": "create_subscription_api_v1_strategies__strategy_code__subscriptions_post",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_code",
            "required": true,
            "schema": {
              "title": "Strategy Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Subscription Api V1 Strategies  Strategy Code  Subscriptions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Subscription",
        "tags": [
          "strategies"
        ]
      }
    },
    "/api/v1/subscriptions/": {
      "get": {
        "operationId": "list_subscriptions_api_v1_subscriptions__get",
        "parameters": [
          {
            "in": "query",
            "name": "strategy_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strategy Code"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {},
                  "title": "Response List Subscriptions Api V1 Subscriptions  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Subscriptions",
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/{sub_id}": {
      "delete": {
        "operationId": "remove_subscription_api_v1_subscriptions__sub_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "sub_id",
            "required": true,
            "schema": {
              "title": "Sub Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Subscription",
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/{sub_id}/backtest": {
      "get": {
        "description": "Return the cached backtest result for a subscription (200) or 404 if never run.",
        "operationId": "get_subscription_backtest_api_v1_subscriptions__sub_id__backtest_get",
        "parameters": [
          {
            "in": "path",
            "name": "sub_id",
            "required": true,
            "schema": {
              "title": "Sub Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Subscription Backtest Api V1 Subscriptions  Sub Id  Backtest Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Subscription Backtest",
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/{sub_id}/positions": {
      "get": {
        "operationId": "get_subscription_positions_api_v1_subscriptions__sub_id__positions_get",
        "parameters": [
          {
            "in": "path",
            "name": "sub_id",
            "required": true,
            "schema": {
              "title": "Sub Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Subscription Positions Api V1 Subscriptions  Sub Id  Positions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Subscription Positions",
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/{sub_id}/start": {
      "post": {
        "operationId": "start_subscription_api_v1_subscriptions__sub_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "sub_id",
            "required": true,
            "schema": {
              "title": "Sub Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Start Subscription Api V1 Subscriptions  Sub Id  Start Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Subscription",
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/{sub_id}/stop": {
      "post": {
        "operationId": "stop_subscription_api_v1_subscriptions__sub_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "sub_id",
            "required": true,
            "schema": {
              "title": "Sub Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Stop Subscription Api V1 Subscriptions  Sub Id  Stop Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop Subscription",
        "tags": [
          "subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/{sub_id}/trades": {
      "get": {
        "operationId": "get_subscription_trades_api_v1_subscriptions__sub_id__trades_get",
        "parameters": [
          {
            "in": "path",
            "name": "sub_id",
            "required": true,
            "schema": {
              "title": "Sub Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Subscription Trades Api V1 Subscriptions  Sub Id  Trades Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Subscription Trades",
        "tags": [
          "subscriptions"
        ]
      }
    },
...
  }
}